refactor(BE): コード整理
