+++ packages/backend/src/controller/index.ts
@@ -1,3 +1,2 @@
-export { register, login, logout, me } from "./auth.controller.js"
-export { session } from "./user.controller.js"
+export { createUser, session } from "./auth.controller.js"
 export { createProject, getProjects, updateProject, deleteProject } from "./project.controller.js"
